NjaxCopilot-Crowned-Edition-20250630_003927/requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0
PyMuPDF==1.23.8
openai==1.3.7
Pillow==10.0.1
pytesseract==0.3.10
requests
pytest
bandit
safety
opencv-python==4.8.1.78
rembg==2.0.50
numpy==1.24.3
scikit-image==0.21.0
termcolor
